# frontend/src/types/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/features/{id}":{"get":{"tags":["feature-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FeatureResponse"}}}}}},"put":{"tags":["feature-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FeatureResponse"}}}}}},"delete":{"tags":["feature-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/features":{"get":{"tags":["feature-controller"],"operationId":"list","parameters":[{"name":"bbox","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FeatureResponse"}}}}}}},"post":{"tags":["feature-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FeatureResponse"}}}}}}},"/api/features/{featureId}/categories":{"get":{"tags":["feature-category-controller"],"operationId":"list_1","parameters":[{"name":"featureId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponse"}}}}}}},"post":{"tags":["feature-category-controller"],"operationId":"associate","parameters":[{"name":"featureId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/features/{featureId}/attachments":{"get":{"tags":["attachment-controller"],"operationId":"list_2","parameters":[{"name":"featureId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AttachmentResponse"}}}}}}},"post":{"tags":["attachment-controller"],"operationId":"upload","parameters":[{"name":"featureId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AttachmentResponse"}}}}}}},"/api/categories":{"get":{"tags":["category-controller"],"operationId":"list_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponse"}}}}}}},"post":{"tags":["category-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Category"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryResponse"}}}}}}},"/api/features/{featureId}/attachments/{attachmentId}":{"delete":{"tags":["attachment-controller"],"operationId":"delete_1","parameters":[{"name":"featureId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"attachmentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/categories/{id}":{"delete":{"tags":["category-controller"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"FeatureRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"featureType":{"type":"string","minLength":1},"description":{"type":"string"},"geom":{"type":"object","additionalProperties":{"type":"object"}},"properties":{"type":"object","additionalProperties":{"type":"object"}}},"required":["geom"]},"FeatureResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"featureType":{"type":"string"},"description":{"type":"string"},"geom":{"type":"object","additionalProperties":{"type":"object"}},"area_m2":{"type":"number","format":"double"},"perimeter_m":{"type":"number","format":"double"},"length_m":{"type":"number","format":"double"},"elevation_min_m":{"type":"number","format":"double"},"elevation_max_m":{"type":"number","format":"double"},"elevation_avg_m":{"type":"number","format":"double"},"properties":{"type":"object","additionalProperties":{"type":"object"}}}},"AttachmentResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"fileName":{"type":"string"},"mimeType":{"type":"string"},"fileSizeBytes":{"type":"integer","format":"int64"},"createdAt":{"type":"string"}}},"Category":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"color":{"type":"string"},"icon":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"CategoryResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"color":{"type":"string"},"icon":{"type":"string"}}}}}}